rename type definition file from `transactions.ts` to `transactions.d.ts`
